Use CAPWAP-AC with HTTP-API
CAPWAP-AC got an upgrade which provides a simple HTTP API to take over
the functionality of the now not-existent CLI.

Note: The support library *jsx* (a JSON Encoder/Decoder) needs special
treatment in rebar.config due to subtle behaviour of rebar3 which
deletes the result of certain compile time checks of *jsx* resulting
in *jsx* being not able to encode / decode Erlang maps. See
https://github.com/erlang/rebar3/issues/1360#issuecomment-256929014
for more details.