Fix building Travis-CI based Docker images
The built container images falling out of Travis seem to be broken: They
seem to contain artifacts of another build-run. As it seems, building
the debian packages also happens via Docker: the current directory is
volume-mapped into the build-container. As part of that, it leaves a
_build directory.

This commit first builds the pure docker image to start from a green
field.

In addition, a recent version of `docker` is used: For whatever reason,
the resulting container image does not contain any .beam files which
is a quite unfortunate situation ...